rename nabto_device_event_handler to nabto_device_listener and fix the nabto_device_authorization_events code,
