+++ src/api/nabto_device.c
@@ -1,11 +1,10 @@
 #include <api/nabto_device_authorization.h>
 #include <api/nabto_device_coap.h>
-#include <api/nabto_device_coap.h>
 #include <api/nabto_device_defines.h>
 #include <api/nabto_device_error.h>
-#include <api/nabto_device_event_handler.h>
 #include <api/nabto_device_future.h>
 #include <api/nabto_device_future_queue.h>
+#include <api/nabto_device_listener.h>
 #include <api/nabto_device_logging.h>
 #include <api/nabto_device_platform.h>
 #include <api/nabto_device_stream.h>
