+++ src/api/nabto_device_authorization_events.c
@@ -3,7 +3,7 @@
 #include <api/nabto_device_authorization.h>
 #include <api/nabto_device_defines.h>
 #include <api/nabto_device_error.h>
-#include <api/nabto_device_event_handler.h>
+#include <api/nabto_device_listener.h>
 
 #include <platform/np_error_code.h>
 
@@ -13,6 +13,27 @@ struct nabto_device_listen_new_authorization_request_context {
     NabtoDeviceEvent* userEvent;
 };
 
+/**
+ * The authorization listener callback is called from the nabto_device_listener
+ * implementation. When an authorization request is ready to be dispatched to a
+ * future or when the listener is closed or an error condition occurs, this
+ * function is called.
+ *
+ * @param ec  NABTO_EC_OK if an event in eventData is ready to be
+ *            NABTO_EC_STOPPED if the listener has been stopped.
+ *            NABTO_EC_ABORTED after all events has been processed and the
+ *            listener is stopped/aborted and ready to be freed.
+ * @param future The future which is going to be resolved after this function
+ *               returns.
+ * @param eventData is a struct nabto_device_authorization_request* which
+ *                  contains an authorization request which comes from the core.
+ * @param listenerData is the structure linked to the NabtoDeviceListener which
+ *                     in this context is a struct
+ *                     nabto_device_authorization_module.
+ *
+ * @return In case the function is called with NABTO_EC_OK, the return value is
+ *         the errorcode the future is resolved with.
+ */
 np_error_code nabto_device_authorization_listener_callback(const np_error_code ec, struct nabto_device_future* future, void* eventData, void* listenerData)
 {
     (void)future;
@@ -23,17 +44,15 @@ np_error_code nabto_device_authorization_listener_callback(const np_error_code e
         struct nabto_device_authorization_request* authReq = (struct nabto_device_authorization_request*)eventData;
         nabto_device_authorization_request_ref_inc(authReq);
         *ctx->request = (NabtoDeviceAuthorizationRequest*)authReq;
-    } else if (ec == NABTO_EC_OUT_OF_MEMORY) {
-        // new auth request cannot be added to the auth queue.
-    } else if (ec == NABTO_EC_ABORTED) {
-        // listener has been freed.
-    } else if (ec == NABTO_EC_STOPPED) {
-        // listener has been stopped.
+    } else if (eventData != NULL) {
+        // We are handling some error, if eventData is present do a false verdict.
+        struct nabto_device_authorization_request* authReq = (struct nabto_device_authorization_request*)eventData;
+        nabto_device_authorization_do_verdict(authReq, false);
     }
     return retEc;
 }
 
-NABTO_DEVICE_DECL_PREFIX NabtoDeviceError  NABTO_DEVICE_API
+NABTO_DEVICE_DECL_PREFIX NabtoDeviceError NABTO_DEVICE_API
 nabto_device_authorization_request_init_listener(NabtoDevice* device, NabtoDeviceListener* authorizationListener)
 {
     struct nabto_device_context* dev = (struct nabto_device_context*)device;
