+++ src/api/nabto_device_listener.h
