+++ src/api/nabto_device_stream.c
@@ -1,6 +1,6 @@
 #include "nabto_device_stream.h"
-#include "nabto_device_event_handler.h"
 #include "nabto_device_future.h"
+#include "nabto_device_listener.h"
 
 #include <api/nabto_device_defines.h>
 #include <api/nabto_device_error.h>
